+++ tortoise/backends/sqlite/schema_generator.py
@@ -1,16 +1,8 @@
-from typing import Optional
-
-from tortoise import fields
 from tortoise.backends.base.schema_generator import BaseSchemaGenerator
 
 
 class SqliteSchemaGenerator(BaseSchemaGenerator):
-    FIELD_TYPE_MAP = {
-        **BaseSchemaGenerator.FIELD_TYPE_MAP,
-        fields.BooleanField: "INTEGER",
-        fields.FloatField: "REAL",
-        fields.DecimalField: "VARCHAR(40)",
-    }
+    DIALECT = "sqlite"
 
     def _escape_comment(self, comment: str) -> str:  # pylint: disable=R0201
         # This method provides a default method to escape comment strings as per
@@ -24,13 +16,6 @@ class SqliteSchemaGenerator(BaseSchemaGenerator):
         _escape_table[ord("\032")] = "\\Z"
         _escape_table[ord("/")] = "\\/"
         return comment.translate(_escape_table)
-
-    def _get_primary_key_create_string(
-        self, field_object: fields.Field, field_name: str, comment: str
-    ) -> Optional[str]:
-        if isinstance(field_object, (fields.SmallIntField, fields.IntField, fields.BigIntField)):
-            return f'"{field_name}" INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL{comment}'
-        return None
 
     def _table_comment_generator(self, table: str, comment: str) -> str:
         return f" /* {self._escape_comment(comment)} */"