+++ tortoise/query_utils.py
@@ -4,8 +4,8 @@ from typing import Any, Dict, List, Optional, Tuple
 from pypika import Table
 from pypika.terms import Criterion
 
-from tortoise import fields
 from tortoise.exceptions import FieldError, OperationalError
+from tortoise.fields.relational import BackwardFKRelation, ManyToManyFieldInstance
 
 
 def _process_filter_kwarg(model, key, value) -> Tuple[Criterion, Optional[Tuple[Table, Criterion]]]:
@@ -44,10 +44,10 @@ def _get_joins_for_related_field(
     required_joins = []
 
     table_pk = related_field.model._meta.db_pk_field
-    related_table_pk = related_field.field_type._meta.db_pk_field
+    related_table_pk = related_field.model_class._meta.db_pk_field
 
-    if isinstance(related_field, fields.ManyToManyFieldInstance):
-        related_table = related_field.field_type._meta.basetable
+    if isinstance(related_field, ManyToManyFieldInstance):
+        related_table = related_field.model_class._meta.basetable
         through_table = Table(related_field.through)
         required_joins.append(
             (
@@ -62,8 +62,8 @@ def _get_joins_for_related_field(
                 == getattr(related_table, related_table_pk),
             )
         )
-    elif isinstance(related_field, fields.BackwardFKRelation):
-        related_table = related_field.field_type._meta.basetable
+    elif isinstance(related_field, BackwardFKRelation):
+        related_table = related_field.model_class._meta.basetable
         required_joins.append(
             (
                 related_table,
@@ -71,7 +71,7 @@ def _get_joins_for_related_field(
             )
         )
     else:
-        related_table = related_field.field_type._meta.basetable
+        related_table = related_field.model_class._meta.basetable
         required_joins.append(
             (
                 related_table,
@@ -213,7 +213,7 @@ class Q:
         related_field = model._meta.fields_map[related_field_name]
         required_joins = _get_joins_for_related_field(table, related_field, related_field_name)
         modifier = Q(**{"__".join(key.split("__")[1:]): value}).resolve(
-            model=related_field.field_type,
+            model=related_field.model_class,
             annotations=self._annotations,
             custom_filters=self._custom_filters,
         )